Add backup retention with `forget`
After running a backup, `restic forget` is called to remove snapshots
that are no longer needed.  "No longer needed" is controlled by setting
a number of hourly/daily/weekly/monthly/yearly snapshots to retain.
This configuration is via environment variables, so check that those
environment variables are set.

NOTE: When `restic backup` is run, it does not explicitly mark a
snapshot as an "hourly", "daily", etc. type of snapshot.  Instead, when
`restic forget` is run, all the snapshots are checked; and assigned an
hourly, daily, etc. label; based on when they were taken.

WARNING: This just deletes snapshots, it does not actually delete pack
files.  That needs to be done separately, but until an unused pack file
is pruned, all runs of `restic check` will complain.