After forget, run prune
Thus runs `restic prune`, to remove pack files that are no longer
needed.  However, some wasted space is allowed: The `REPACK_MAX_UNUSED`
environment variable takes a percentage or a number of bytes (with or
without k/m/g suffix); packs are allowed to have that much space unused
before a repack is triggered.

Setting a non-zero number can be good when using a cloud storage
platform that charges for data retrieval, or that charges more for
regularly accessing files (instead of letting them age).